--- a/src/clj/rems/markdown.clj
+++ b/src/clj/rems/markdown.clj
@@ -89,3 +89,40 @@
                                 "```"
                                 "the end"]))
       "keeps Clojure code intact"))
+
+(defn- apply-padding [coll]
+  (let [max-len (apply max (map count coll))
+        get-padding #(repeat (- max-len (count %)) " ")]
+    (->> coll
+         (mapv #(apply str % (get-padding %))))))
+
+(defn- add-table-header-separator [coll]
+  (let [make-separator (comp str/join #(repeat (count %) "-"))]
+    (vec
+     (concat (take 1 coll)
+             (conj [] (mapv make-separator (first coll)))
+             (drop 1 coll)))))
+
+(defn markdown-table [{:keys [header rows row-fn]}]
+  (let [transpose #(apply mapv vector %)
+        format-rows #(->> (str/join " | " %)
+                          (format "| %s |"))
+        row-fn (or row-fn identity)
+        cols (mapv str header)
+        rows (vec (for [row rows]
+                    (mapv str (row-fn row))))]
+    (->> (into [cols] rows)
+         transpose
+         (mapv apply-padding) ; align column width
+         transpose
+         add-table-header-separator
+         (mapv format-rows))))
+
+(comment
+  (doseq [row (markdown-table {:header ["handler count" "mean time" "lower-q 2.5%" "upper-q 97.5%"]
+                               :rows [[1 0.049 0.039 0.067]
+                                      [10 0.423 0.401 0.452]
+                                      [50 2.098 1.910 2.294]
+                                      [100 4.145 3.849 4.673]
+                                      [200 8.539 8.025 9.374]]})]
+    (println row)))