--- a/src/cljc/rems/text.cljc
+++ b/src/cljc/rems/text.cljc
@@ -1,6 +1,5 @@
 (ns rems.text
-  (:require [better-cond.core :as b]
-            [clojure.string :as str]
+  (:require [clojure.string :as str]
             [clojure.test :refer [deftest is]]
             #?(:clj [clj-time.core :as time]
                :cljs [cljs-time.core :as time])
@@ -27,19 +26,31 @@
                  (first args)
                  (cons k args)))))
 
+(def cached-tr (atom nil))
+(defn reset-cached-tr! [] (reset! cached-tr nil))
+
+(defn ensure-cached-tr! []
+  (when-not (fn? @cached-tr)
+    (reset! cached-tr (rems.tempura/get-cached-tr #?(:clj rems.locales/translations
+                                                     :cljs @(rf/subscribe [:translations])))))
+  @cached-tr)
+
+(defn- tr [ks & [args]]
+  (let [language #?(:clj context/*lang*
+                    :cljs @(rf/subscribe [:language]))]
+    ((ensure-cached-tr!) [language]
+                         (vec ks)
+                         (some-> args vec))))
+
 (defn text-format
   "Return the tempura translation for a given key and arguments:
 
    `(text-format :key 1 2)`"
   [k & args]
-  #?(:clj (rems.tempura/tr rems.locales/translations
-                           context/*lang*
-                           [k :t/missing]
-                           args)
-     :cljs (rems.tempura/tr @(rf/subscribe [:translations])
-                            @(rf/subscribe [:language])
-                            [k :t/missing (failsafe-fallback k args)]
-                            args)))
+  #?(:clj (tr [k :t/missing]
+              args)
+     :cljs (tr [k :t/missing (failsafe-fallback k args)]
+               args)))
 
 (defn text-format-map
   "Return the tempura translation for a given key and argument map:
@@ -58,13 +69,9 @@
   "Return the tempura translation for a given key. Additional fallback
   keys can be given but there is no default fallback text."
   [& ks]
-  #?(:clj (rems.tempura/tr rems.locales/translations
-                           context/*lang*
-                           ks)
+  #?(:clj (tr ks)
      :cljs (try
-             (rems.tempura/tr @(rf/subscribe [:translations])
-                              @(rf/subscribe [:language])
-                              ks)
+             (tr ks)
              (catch js/Object e
                ;; fail gracefully if the re-frame state is incomplete
                (.error js/console e)
@@ -74,15 +81,11 @@
   "Return the tempura translation for a given key. Additional fallback
   keys can be given."
   [& ks]
-  #?(:clj (rems.tempura/tr rems.locales/translations
-                           context/*lang*
-                           (conj (vec ks) (text-format :t/missing (vec ks))))
+  #?(:clj (tr (conj (vec ks) (text-format :t/missing (vec ks))))
      ;; NB: we can't call the text-no-fallback here as in CLJS
      ;; we can both call this as function or use as a React component
      :cljs (try
-             (rems.tempura/tr @(rf/subscribe [:translations])
-                              @(rf/subscribe [:language])
-                              (conj (vec ks) (text-format :t/missing (vec ks))))
+             (tr (conj (vec ks) (text-format :t/missing (vec ks))))
              (catch js/Object e
                ;; fail gracefully if the re-frame state is incomplete
                (.error js/console e)